Removing the dead code for HTTP trailer read.
The public APIs did not provide a way to read the HTTP trailers but the incomplete code existed to read it.
