--- rxnetty-http/src/main/java/io/reactivex/netty/protocol/http/internal/HttpTrailerSubscriberEvent.java
@@ -1,33 +1,0 @@
-/*
- * Copyright 2015 Netflix, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- */
-package io.reactivex.netty.protocol.http.internal;
-
-import io.reactivex.netty.protocol.http.TrailingHeaders;
-import rx.Subscriber;
-
-public class HttpTrailerSubscriberEvent {
-
-    private final Subscriber<? super TrailingHeaders> subscriber;
-
-    public HttpTrailerSubscriberEvent(Subscriber<? super TrailingHeaders> subscriber) {
-        this.subscriber = subscriber;
-    }
-
-    public Subscriber<? super TrailingHeaders> getSubscriber() {
-        return subscriber;
-    }
-}
